requirements.txt
python-dotenv==0.19.0
tweepy==3.9.0

